Relocate modules into 'octrees' directory, and update associated import references
