+++ octrees/geometry.py
